+++ CHANGELOG.md
@@ -29,6 +29,9 @@ Breaking changes can occur in Minor releases.
 - Added support for different monitoring backends
     - Perf-based : Default, uses the Linux `perf` tool to collect performance metrics
     - Daemon-based : Uses a custom daemon to collect performance metrics.
+- Added support for thermal monitoring of cores on Intel CPUs (using either `coretemp` or accessing the msr directly)
+    - Updated postprocessing library to parse core temperatures from the periodic log
+    - Added new clip which visualizes the core temperatures over time
 
 ### Changed 
 - `Scheduler` exposes a new method `is_time_to_launch(application, system_state)`, which allows to make scheduling decisions based on the current system state.